Created model base
