+++ src/helpers/orbitdb/MeetupModel.ts
@@ -1,0 +1,4 @@
+import ModelBase from './ModelBase'
+import { IMeetupInfo } from '../../interfaces';
+
+export default class PostModel extends ModelBase<IMeetupInfo> { }
