+++ src/interfaces.ts
@@ -1,5 +1,5 @@
 interface IContentInfo {
-    _id: string;
+    _id?: string;
     title: string;
     author: string;
     date: number;
